refactor: add types file; create function to display coordinates in map
