+++ src/types.ts
@@ -1,0 +1,12 @@
+export interface ICoordinates {
+    lat: {
+        degrees: number;
+        minutes: number;
+        seconds: number;
+    };
+    lon: {
+        degrees: number;
+        minutes: number;
+        seconds: number;
+    };
+}
